requirements.txt
groq
langchain
langchain_core
langchain_groq
newspaper3k
streamlit
lxml[html_clean]